+++ src-tauri/src/dirstat/mod.rs
@@ -13,6 +13,8 @@ use ts_rs::TS;
 #[cfg(target_os = "windows")]
 use winapi_util::{file, Handle};
 
+use crate::utils::get_percentage;
+
 static INIT: Once = Once::new();
 static mut COUNTER: Option<AtomicU64> = None;
 
@@ -38,6 +40,8 @@ pub struct DiskItem {
     pub children: Option<Vec<DiskItem>>,
     #[ts(type = "number")]
     pub size: u64,
+    #[ts(type = "number")]
+    pub percentage_of_disk: f64,
 }
 
 // Define a type alias for the callback
@@ -116,6 +120,7 @@ impl DiskItem {
     }
 
     pub fn scan(
+        is_turbo: &bool,
         path: &Path,
         apparent: bool,
         root_dev: u64,
@@ -142,29 +147,53 @@ impl DiskItem {
                     .filter_map(Result::ok)
                     .collect::<Vec<_>>();
 
-                let sub_items: Vec<DiskItem> = sub_entries
-                    .into_iter()
-                    .filter_map(|entry| {
-                        let result = DiskItem::scan(
-                            &entry.path(),
-                            apparent,
-                            root_dev,
-                            callback,
-                            total_bytes,
-                            Arc::clone(&bytes_scanned),
-                        );
+                let sub_items: Vec<DiskItem> = match is_turbo {
+                    // Use rayon parallel iterator if turbo
+                    true => sub_entries
+                        .into_par_iter()
+                        .filter_map(|entry| {
+                            let result = DiskItem::scan(
+                                is_turbo,
+                                &entry.path(),
+                                apparent,
+                                root_dev,
+                                callback,
+                                total_bytes,
+                                Arc::clone(&bytes_scanned),
+                            );
 
-                        result.ok()
-                    })
-                    .collect();
+                            result.ok()
+                        })
+                        .collect(),
+                    // Use normal iter
+                    false => sub_entries
+                        .into_iter()
+                        .filter_map(|entry| {
+                            let result = DiskItem::scan(
+                                is_turbo,
+                                &entry.path(),
+                                apparent,
+                                root_dev,
+                                callback,
+                                total_bytes,
+                                Arc::clone(&bytes_scanned),
+                            );
+
+                            result.ok()
+                        })
+                        .collect(),
+                };
 
                 let mut sorted_sub_items = sub_items;
                 sorted_sub_items.sort_unstable_by(|a, b| a.size.cmp(&b.size).reverse());
 
+                let size = sorted_sub_items.iter().map(|item| item.size).sum();
+
                 Ok(DiskItem {
                     id,
                     name,
-                    size: sorted_sub_items.iter().map(|item| item.size).sum(),
+                    size,
+                    percentage_of_disk: get_percentage(&size, &total_bytes),
                     children: Some(sorted_sub_items),
                 })
             }
@@ -176,72 +205,7 @@ impl DiskItem {
                     name,
                     size,
                     children: None,
-                })
-            }
-        }
-    }
-
-    pub fn turbo_scan(
-        path: &Path,
-        apparent: bool,
-        root_dev: u64,
-        callback: &ProgressCallback,
-        total_bytes: u64,
-        bytes_scanned: Arc<AtomicU64>,
-    ) -> Result<Self, Box<dyn Error>> {
-        let id = get_next_id();
-        let name = path
-            .file_name()
-            .unwrap_or(OsStr::new("."))
-            .to_string_lossy()
-            .to_string();
-
-        let file_info = FileInfo::from_path(path, apparent)?;
-
-        match file_info {
-            FileInfo::Directory { volume_id } => {
-                if volume_id != root_dev {
-                    return Err("Filesystem boundary crossed".into());
-                }
-
-                let sub_entries = fs::read_dir(path)?
-                    .filter_map(Result::ok)
-                    .collect::<Vec<_>>();
-
-                let sub_items: Vec<DiskItem> = sub_entries
-                    .into_par_iter()
-                    .filter_map(|entry| {
-                        let result = DiskItem::turbo_scan(
-                            &entry.path(),
-                            apparent,
-                            root_dev,
-                            callback,
-                            total_bytes,
-                            Arc::clone(&bytes_scanned),
-                        );
-
-                        result.ok()
-                    })
-                    .collect();
-
-                let mut sorted_sub_items = sub_items;
-                sorted_sub_items.sort_unstable_by(|a, b| a.size.cmp(&b.size).reverse());
-
-                Ok(DiskItem {
-                    id,
-                    name,
-                    size: sorted_sub_items.iter().map(|item| item.size).sum(),
-                    children: Some(sorted_sub_items),
-                })
-            }
-            FileInfo::File { size, .. } => {
-                bytes_scanned.fetch_add(size, Ordering::SeqCst);
-                callback(bytes_scanned.load(Ordering::SeqCst), total_bytes);
-                Ok(DiskItem {
-                    id,
-                    name,
-                    size,
-                    children: None,
+                    percentage_of_disk: get_percentage(&size, &total_bytes),
                 })
             }
         }