+++ CHANGELOG.md
@@ -10,6 +10,11 @@ All notable changes to this project will be documented in this file. This projec
 - Implement logger for better debugging
 - Implement better product analytics
 - Implement better icons for the app
+- Cleanup of Disk analysis code
+- Improved UI of disk analysis page
+- Added progress bar for File explorer
+- Removed Apexchart in favor of Highcharts
+- Add Performance mode to settings
 
 ## 0.9.9
 - Fix issue where some popover components are transparent